Document and reorganize everything.
