+++ host/examples/cpal/Cargo.toml
@@ -5,7 +5,7 @@ edition = "2021"
 
 [dependencies]
 clack-host = { path = "../.." }
-clack-extensions = { path = "../../../extensions", features = ["clack-host", "audio-ports", "audio-ports-config", "gui", "log", "params", "timer"] }
+clack-extensions = { path = "../../../extensions", features = ["clack-host", "audio-ports", "gui", "log", "params", "timer"] }
 cpal = "0.15.2"
 crossbeam-channel = "0.5.8"
 clap = { version = "4.3.0", features = ["derive"] }
